Convert inline CSS when create init model
